More reorgnization for cadl-azure repo
