+++ packages/compiler/core/helpers/index.ts
@@ -1,3 +1,3 @@
+export * from "./operation-utils.js";
 export * from "./projected-names-utils.js";
 export * from "./usage-resolver.js";
-export * from "./operation-utils.js";
