+++ packages/rest/src/http/index.ts
@@ -2,5 +2,6 @@ export const namespace = "Cadl.Http";
 
 export * from "./decorators.js";
 export * from "./operations.js";
+export * from "./parameters.js";
 export * from "./responses.js";
 export * from "./types.js";
